@@ -2,7 +2,7 @@ from setuptools import setup
 
 setup(
     name="jupyterhub-ldapauthenticator",
-    version="1.3.3.dev",
+    version="2.0.0.dev",
     description="LDAP Authenticator for JupyterHub",
     long_description=open("README.md").read(),
     long_description_content_type="text/markdown",
@@ -11,9 +11,9 @@ setup(
     author_email="yuvipanda@riseup.net",
     license="3 Clause BSD",
     packages=["ldapauthenticator"],
-    python_requires=">=3.7",
+    python_requires=">=3.9",
     install_requires=[
-        "jupyterhub",
+        "jupyterhub>=4.1.6",
         "ldap3",
         "tornado",
         "traitlets",
@@ -23,6 +23,12 @@ setup(
             "pytest",
             "pytest-asyncio",
             "pytest-cov",
+        ],
+    },
+    entry_points={
+        "jupyterhub.authenticators": [
+            "ldap = ldapauthenticator:LDAPAuthenticator",
+            "ldapauthenticator = ldapauthenticator:LDAPAuthenticator",
         ],
     },
 )
