@@ -4,7 +4,7 @@ import ldap3
 from jupyterhub.auth import Authenticator
 from ldap3.utils.conv import escape_filter_chars
 from tornado import gen
-from traitlets import Bool, Int, List, Unicode, Union
+from traitlets import Bool, Int, List, Unicode, Union, validate
 
 
 class LDAPAuthenticator(Authenticator):
@@ -64,6 +64,19 @@ class LDAPAuthenticator(Authenticator):
         	]
         """,
     )
+
+    @validate("bind_dn_template")
+    def _validate_bind_dn_template(self, proposal):
+        """
+        Ensure a List[str] is set, filtered from empty string elements.
+        """
+        rv = []
+        if isinstance(proposal.value, str):
+            rv = [proposal.value]
+        if "" in rv:
+            self.log.warning("Ignoring blank 'bind_dn_template' entry!")
+            rv = [e for e in rv if e]
+        return rv
 
     allowed_groups = List(
         config=True,
@@ -131,7 +144,7 @@ class LDAPAuthenticator(Authenticator):
         c.LDAPAuthenticator.lookup_dn_search_user = 'ldap_search_user_technical_account'
         c.LDAPAuthenticator.lookup_dn_search_password = 'secret'
         c.LDAPAuthenticator.user_search_base = 'ou=people,dc=wikimedia,dc=org'
-        c.LDAPAuthenticator.user_attribute = 'sAMAccountName'
+        c.LDAPAuthenticator.user_attribute = 'uid'
         c.LDAPAuthenticator.lookup_dn_user_dn_attribute = 'cn'
         c.LDAPAuthenticator.bind_dn_template = '{username}'
         ```
@@ -226,36 +239,32 @@ class LDAPAuthenticator(Authenticator):
     )
 
     def resolve_username(self, username_supplied_by_user):
+        """
+        Resolves a username supplied by a user to the a user DN when lookup_dn
+        is True.
+        """
         search_dn = self.lookup_dn_search_user
         if self.escape_userdn:
             search_dn = escape_filter_chars(search_dn)
         conn = self.get_connection(
-            userdn=search_dn, password=self.lookup_dn_search_password
+            userdn=search_dn,
+            password=self.lookup_dn_search_password,
         )
-        is_bound = conn.bind()
-        if not is_bound:
-            msg = "Failed to connect to LDAP server with search user '{search_dn}'"
-            self.log.warning(msg.format(search_dn=search_dn))
+        if not conn.bind():
+            self.log.warning(
+                f"Failed to connect to LDAP server with search user '{search_dn}'"
+            )
             return (None, None)
 
         search_filter = self.lookup_dn_search_filter.format(
             login_attr=self.user_attribute,
             login=escape_filter_chars(username_supplied_by_user),
         )
-        msg = "\n".join(
-            [
-                "Looking up user with:",
-                "    search_base = '{search_base}'",
-                "    search_filter = '{search_filter}'",
-                "    attributes = '{attributes}'",
-            ]
-        )
         self.log.debug(
-            msg.format(
-                search_base=self.user_search_base,
-                search_filter=search_filter,
-                attributes=self.user_attribute,
-            )
+            "Looking up user with:\n",
+            f"    search_base = '{self.user_search_base}'\n",
+            f"    search_filter = '{search_filter}'\n",
+            f"    attributes = '[{self.lookup_dn_user_dn_attribute}]'",
         )
         conn.search(
             search_base=self.user_search_base,
@@ -265,14 +274,9 @@ class LDAPAuthenticator(Authenticator):
         )
         response = conn.response
         if len(response) == 0 or "attributes" not in response[0].keys():
-            msg = (
-                "No entry found for user '{username}' "
-                "when looking up attribute '{attribute}'"
-            )
             self.log.warning(
-                msg.format(
-                    username=username_supplied_by_user, attribute=self.user_attribute
-                )
+                f"No entry found for user '{username_supplied_by_user}' "
+                f"when looking up attribute '{self.user_attribute}'"
             )
             return (None, None)
 
@@ -283,19 +287,11 @@ class LDAPAuthenticator(Authenticator):
             elif len(user_dn) == 1:
                 user_dn = user_dn[0]
             else:
-                msg = (
-                    "A lookup of the username '{username}' returned a list "
-                    "of entries for the attribute '{attribute}'. Only the "
-                    "first among these ('{first_entry}') was used. The other "
-                    "entries ({other_entries}) were ignored."
-                )
                 self.log.warn(
-                    msg.format(
-                        username=username_supplied_by_user,
-                        attribute=self.lookup_dn_user_dn_attribute,
-                        first_entry=user_dn[0],
-                        other_entries=", ".join(user_dn[1:]),
-                    )
+                    f"A lookup of the username '{username_supplied_by_user}' returned a list "
+                    f"of entries for the attribute '{self.lookup_dn_user_dn_attribute}'. Only "
+                    f"the first among these ('{user_dn[0]}') was used. The other entries "
+                    f"({', '.join(user_dn[1:])}) were ignored."
                 )
                 user_dn = user_dn[0]
 
@@ -325,6 +321,14 @@ class LDAPAuthenticator(Authenticator):
 
     @gen.coroutine
     def authenticate(self, handler, data):
+        """
+        Note: This function is really meant to identify a user, and
+              check_allowed and check_blocked are meant to determine if its an
+              authorized user. Authorization is currently handled by returning
+              None here instead.
+
+        ref: https://jupyterhub.readthedocs.io/en/latest/reference/authenticators.html#authenticator-authenticate
+        """
         username = data["username"]
         password = data["password"]
 
@@ -342,18 +346,14 @@ class LDAPAuthenticator(Authenticator):
             self.log.warning("username:%s Login denied for blank password", username)
             return None
 
-        # bind_dn_template should be of type List[str]
-        bind_dn_template = self.bind_dn_template
-        if isinstance(bind_dn_template, str):
-            bind_dn_template = [bind_dn_template]
-
         # sanity check
-        if not self.lookup_dn and not bind_dn_template:
+        if not self.lookup_dn and not self.bind_dn_template:
             self.log.warning(
                 "Login not allowed, please configure 'lookup_dn' or 'bind_dn_template'."
             )
             return None
 
+        bind_dn_template = self.bind_dn_template
         if self.lookup_dn:
             username, resolved_dn = self.resolve_username(username)
             if not username:
@@ -366,14 +366,10 @@ class LDAPAuthenticator(Authenticator):
 
         is_bound = False
         for dn in bind_dn_template:
-            if not dn:
-                self.log.warning("Ignoring blank 'bind_dn_template' entry!")
-                continue
             userdn = dn.format(username=username)
             if self.escape_userdn:
                 userdn = escape_filter_chars(userdn)
-            msg = "Attempting to bind {username} with {userdn}"
-            self.log.debug(msg.format(username=username, userdn=userdn))
+            self.log.debug(f"Attempting to bind {username} with {userdn}")
             msg = "Status of user bind {username} with {userdn} : {is_bound}"
             try:
                 conn = self.get_connection(userdn, password)
@@ -391,8 +387,7 @@ class LDAPAuthenticator(Authenticator):
                 break
 
         if not is_bound:
-            msg = "Invalid password for user '{username}'"
-            self.log.warning(msg.format(username=username))
+            self.log.warning(f"Invalid password for user '{username}'")
             return None
 
         if self.search_filter:
@@ -407,20 +402,14 @@ class LDAPAuthenticator(Authenticator):
             )
             n_users = len(conn.response)
             if n_users == 0:
-                msg = "User with '{userattr}={username}' not found in directory"
                 self.log.warning(
-                    msg.format(userattr=self.user_attribute, username=username)
+                    f"User with '{self.user_attribute}={username}' not found in directory"
                 )
                 return None
             if n_users > 1:
-                msg = (
-                    "Duplicate users found! "
-                    "{n_users} users found with '{userattr}={username}'"
-                )
                 self.log.warning(
-                    msg.format(
-                        userattr=self.user_attribute, username=username, n_users=n_users
-                    )
+                    "Duplicate users found! {n_users} users found "
+                    f"with '{self.user_attribute}={username}'"
                 )
                 return None
 
@@ -450,8 +439,9 @@ class LDAPAuthenticator(Authenticator):
                     break
             if not found:
                 # If we reach here, then none of the groups matched
-                msg = "username:{username} User not in any of the allowed groups"
-                self.log.warning(msg.format(username=username))
+                self.log.warning(
+                    f"username:{username} User not in any of the allowed groups"
+                )
                 return None
 
         if not self.use_lookup_dn_username:
