@@ -3,4 +3,4 @@ from ldapauthenticator.ldapauthenticator import LDAPAuthenticator  # noqa
 # __version__ should be updated using tbump, based on configuration in
 # pyproject.toml, according to instructions in RELEASE.md.
 #
-__version__ = "1.3.3.dev"
+__version__ = "2.0.0.dev"
